+++ comp2013/src/main/java/comp2013/SnakeView.java
@@ -202,8 +202,12 @@ public class SnakeView extends Application implements IView {
         // image accordingly.
         switch (this.m_Controller.m_Snake.getDirection()) {
             case SnakeObject.UP: {
-                M_SnakeHeadImg = SnakeImageUtil.getImage
-                        ("snakeHeadUp");
+                //M_SnakeHeadImg = SnakeImageUtil.getImage
+                       // ("snakeHeadUp");
+                M_SnakeHeadImg = SnakeImageUtil.getSprite
+                        ("headUp", false)
+                        .snapshot(null,
+                                null);
             }
             break;
             case SnakeObject.DOWN: {